[AIRFLOW-1314] major refactor to request factory
